BERTopic working on a testing dataset.
I didn't use it on our dataset so I can compare results online.
It takes forever so I just used 5000 so the topics aren't great but it gives good reason to experiment.
I want to try a better representation with less stop words in it and then try it on our data